refactor: move common generator methods into module
